Huge refactoring - change highlight module structure. Invented TextHighlighter interface
