+++ .gitattributes
@@ -1,0 +1,2 @@
+*.so filter=lfs diff=lfs merge=lfs -text
+*.dll filter=lfs diff=lfs merge=lfs -text
