+++ src/main/java/cdeler/highlight/highlighters/TextHighlighter.java
@@ -1,0 +1,7 @@
+package cdeler.highlight.highlighters;
+
+import javax.swing.*;
+
+public interface TextHighlighter {
+    void highlight(JTextPane textArea);
+}
