+++ src/main/java/cdeler/highlight/token/AST.java
@@ -1,4 +1,4 @@
-package cdeler.highlight;
+package cdeler.highlight.token;
 
 import java.util.List;
 import java.util.function.Function;
