+++ src/main/java/cdeler/highlight/token/AbstractTokenizer.java
@@ -1,10 +1,12 @@
-package cdeler.highlight;
+package cdeler.highlight.token;
 
 import java.util.Collections;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import cdeler.highlight.HighlightException;
 
 public abstract class AbstractTokenizer<T> implements Tokenizer {
     private static final Logger LOGGER = LoggerFactory.getLogger(AbstractTokenizer.class);
