+++ src/main/java/cdeler/highlight/token/SourceToken.java
@@ -1,4 +1,4 @@
-package cdeler.highlight;
+package cdeler.highlight.token;
 
 import java.util.Objects;
