+++ src/main/java/cdeler/highlight/token/Token.java
@@ -1,4 +1,4 @@
-package cdeler.highlight;
+package cdeler.highlight.token;
 
 public interface Token {
