+++ src/main/java/cdeler/highlight/token/Tokenizer.java
@@ -1,4 +1,4 @@
-package cdeler.highlight;
+package cdeler.highlight.token;
 
 import java.util.List;
