+++ src/main/resources/app-context.xml
@@ -3,9 +3,9 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
        http://www.springframework.org/schema/beans/spring-beans-2.5.xsd">
 
-    <bean id="jni_tokenizer" class="cdeler.highlight.JNITokenizer"/>
+    <bean id="jni_tokenizer" class="cdeler.highlight.token.JNITokenizer"/>
 
-    <bean id="highlighter" class="cdeler.highlight.StyleTextHighlighter">
+    <bean id="highlighter" class="cdeler.highlight.highlighters.StyleTextHighlighter">
         <constructor-arg index="0" ref="jni_tokenizer"/>
         <constructor-arg index="1" ref="settings_manager"/>
     </bean>
